Relocated monitor
